Split app.common.message_moving into a submodule.
